pyproject.toml
[project]
name = "repl-nix-dashmanagedb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "numpy==1.24.3",
    "openai>=1.52.2",
    "pandas>=2.1.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
    "requests>=2.32.3",
    "scikit-learn>=1.5.2",
    "selenium==4.15.2",
    "sendgrid>=6.11.0",
    "spacy==3.7.2",
    "streamlit>=1.39.0",
    "trafilatura>=1.12.2",
    "webdriver-manager==4.0.1",
    "werkzeug>=3.0.6",
]
